ai-face-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
deepface==0.0.79
opencv-python==4.8.1.78
numpy==1.24.3
pandas==2.0.3
Pillow==10.0.1
scikit-learn==1.3.0
requests==2.31.0
python-dotenv==1.0.0 